fix build
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@2072 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/battle2/battle_stats.cpp
@@ -744,7 +744,7 @@ u32 Battle2::Stats::CalculateDamageStats(const Stats & enemy, double dmg) const
 	default: break;
     }
 
-    return dmg;
+    return static_cast<u32>(dmg);
 }
 
 u32 Battle2::Stats::GetDamage(const Stats & enemy) const
